[mr2492] fix many compiler warnings by result of static code analyser (part N+1).

+++ src/game/vmap/RegularGrid.h
@@ -86,7 +86,7 @@ public:
 
         static Cell ComputeCell(float fx, float fy)
         {
-            Cell c = {fx * (1.f/CELL_SIZE) + (CELL_NUMBER/2), fy * (1.f/CELL_SIZE) + (CELL_NUMBER/2)};
+            Cell c = {int(fx * (1.f/CELL_SIZE) + ((float)CELL_NUMBER/2.0f)), int(fy * (1.f/CELL_SIZE) + ((float)CELL_NUMBER/2.0f))};
             return c;
         }
